refactor: put create_experiment methods in tracel-core
