+++ crates/tracel-experiment/src/remote/mod.rs
@@ -1,10 +1,6 @@
 pub mod base;
-mod cloud;
+pub mod cloud;
 mod logs;
 mod socket;
 #[cfg(feature = "station")]
 pub mod station;
-
-pub use cloud::create_cloud_experiment_run;
-#[cfg(feature = "station")]
-pub use station::create_station_experiment_run;
